requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
arxiv>=2.1.0
requests>=2.31.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
scikit-learn>=1.4.0
google-genai>=0.7.0
python-multipart>=0.0.9
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
